Move __init__ back to job
